fig: delete image avatar product admin

+++ src/@App/pages/Admin/Product/components/UploadThumbnail.jsx
@@ -47,12 +47,7 @@ function UploadThumbnail({ name, control, multiple = false, sx, title, product_i
                await productService.createImage({ images: res, product_id });
                refetchImages();
             } else {
-               console.log(res);
-               if (typeof res === 'string') {
-                  onChange([res, ...imageOrImages]);
-               } else {
-                  onChange([...res, ...imageOrImages]);
-               }
+               onChange([...res, ...imageOrImages]);
             }
             setIsChangeImages((prev) => !prev);
          } else {
@@ -67,25 +62,19 @@ function UploadThumbnail({ name, control, multiple = false, sx, title, product_i
    const { mutate: callbackDeleteImage, isLoading: deleteLoading } = useMutation({
       mutationKey: 'uploadImage',
       mutationFn: async ({ image, id }) => {
-         if (product_id) {
-            if (multiple) {
-               console.log(image, id);
-               const res = await productService.deleteImage(id);
-               if (res.success) {
-                  await deleteFirebaseImage(image);
-                  setIsChangeImages((prev) => !prev);
-                  refetchImages();
-               } else {
-                  errorMessage('Hình ảnh đang được sử dụng.');
-               }
+         if (product_id && multiple) {
+            const res = await productService.deleteImage(id);
+            if (res.success) {
+               await deleteFirebaseImage(image);
+               setIsChangeImages((prev) => !prev);
+               refetchImages();
+            } else {
+               errorMessage('Hình ảnh đang được sử dụng.');
             }
-         } else {
+         } else if (!multiple) {
             await deleteFirebaseImage(image);
             product_id && (await productService.deleteThumbnail(product_id));
-
-            const newValue = imageOrImages.filter((item) => item !== image);
-
-            onChange(newValue);
+            onChange('');
          }
       },
       onError: (error) => {
@@ -118,14 +107,14 @@ function UploadThumbnail({ name, control, multiple = false, sx, title, product_i
                      imageOrImages.map((image, index) => {
                         return (
                            <ImageItem key={index} xs={sx}>
-                              <LazyLoadingImage src={image.image_url || image} style={{ borderRadius: '5px' }} />
+                              <LazyLoadingImage src={image.image_url || image} style={{ borderRadius: '5px', height: '180px' }} />
                               <DeleteImage
-                                 onClick={() => {
+                                 onClick={() =>
                                     callbackDeleteImage({
                                        image: image.image_url ? image.image_url : image,
                                        id: image?._id
-                                    });
-                                 }}
+                                    })
+                                 }
                               />
                            </ImageItem>
                         );
@@ -166,7 +155,6 @@ function UploadThumbnail({ name, control, multiple = false, sx, title, product_i
 const ImageItem = styled('label')({
    position: 'relative',
    width: '155.227px',
-   height: '178.400px',
    display: 'flex',
    justifyContent: 'center',
    alignItems: 'center',
